+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 * Added method `vdc.QueryEdgeGateway` [#364](https://github.com/vmware/go-vcloud-director/pull/364)
 * Deprecated `vdc.GetEdgeGatewayRecordsType` [#364](https://github.com/vmware/go-vcloud-director/pull/364)
 * Dropped support for VCD 9.7 which is EOL now [#xxx](https://github.com/vmware/go-vcloud-director/pull/xxx)
+* Bumped Default API Version to V33.0  [#xxx](https://github.com/vmware/go-vcloud-director/pull/xxx)
+* Field `types.Disk.Size` is replaced with `types.Disk.SizeMb` as size in Kilobytes is not supported in V33.0  [#xxx](https://github.com/vmware/go-vcloud-director/pull/xxx)
 
 ## 2.11.0 (March 10, 2021)